requirements.txt
streamlit>=1.22
requests>=2.28
python-dotenv>=1.0
gTTS>=2.3
pillow>=9.0
